(string_list) Silence warning

+++ libretro-common/lists/string_list.c
@@ -452,7 +452,7 @@ int string_list_find_elem(const struct string_list *list, const char *elem)
       for (i = 0; i < list->size; i++)
       {
          if (string_is_equal_noncase(list->elems[i].data, elem))
-            return (i + 1);
+            return (int)(i + 1);
       }
    }
    return 0;
